requirements.txt
langchain
langchain_community
python-dotenv
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
fastapi
uvicorn
langserve
sse_starlette
arxiv
wikipedia
langchainhub
langsmith
duckduckgo-search
pypdf